% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{load_fonts}
\alias{load_fonts}
\title{Load \link{wplot} fonts}
\usage{
load_fonts()
}
\description{
The fonts will be loaded into both the \href{https://cran.r-project.org/web/packages/sysfonts/index.html}{\code{sysfonts}} and \code{\link{extrafont}} databases
}
