% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emplot.R
\name{theme_emplot}
\alias{theme_emplot}
\title{Default emplot theme (light)}
\usage{
theme_emplot(base_size = 12, grid = TRUE, serif = FALSE)
}
\arguments{
\item{base_size}{base font size}

\item{grid}{show grid lines}

\item{serif}{use serif font}
}
\value{
theme object
}
\description{
Add to a ggplot2 plot object to apply.
}
