% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emplotdark.R
\name{theme_emplotdark}
\alias{theme_emplotdark}
\title{Emplot dark theme}
\usage{
theme_emplotdark(base_size = 9, grid = TRUE)
}
\arguments{
\item{base_size}{base font size}

\item{grid}{show grid lines}
}
\value{
theme object
}
\description{
Add to a ggplot2 plot object to apply.
}
