% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emtufte.R
\name{theme_emtufte}
\alias{theme_emtufte}
\title{Tufte emplot theme}
\usage{
theme_emtufte(base_size = 10, grid = FALSE, serif = TRUE)
}
\arguments{
\item{base_size}{base font size}

\item{grid}{show grid lines}

\item{serif}{use serif font}
}
\value{
theme object
}
\description{
Simular to the default theme, but with Tufte-esque defaults.
}
