% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_w.R
\name{theme_w}
\alias{theme_w}
\title{Default wplot theme (light)}
\usage{
theme_w(base_size = 9, grid = TRUE, serif = FALSE)
}
\arguments{
\item{base_size}{base font size}

\item{grid}{show grid lines}

\item{serif}{use serif font}
}
\value{
theme object
}
\description{
Add to a ggplot2 plot object to apply.
}
