% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mc.R
\name{theme_mc}
\alias{theme_mc}
\title{mplot classic theme}
\usage{
theme_mc(base_size = 10, grid = FALSE, serif = TRUE)
}
\arguments{
\item{base_size}{base font size}

\item{grid}{show grid lines}

\item{serif}{use serif font}
}
\value{
theme object
}
\description{
Similar to the default theme, but with larger serif font and no gridlines.
}
