% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_md.R
\name{theme_md}
\alias{theme_md}
\title{mplot dark theme}
\usage{
theme_md(base_size = 9, grid = TRUE)
}
\arguments{
\item{base_size}{base font size}

\item{grid}{show grid lines}
}
\value{
theme object
}
\description{
Add to a ggplot2 plot object to apply.
}
