% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mt.R
\name{theme_mt}
\alias{theme_mt}
\title{mplot Tufte theme}
\usage{
theme_mt(base_size = 10, grid = FALSE, serif = TRUE)
}
\arguments{
\item{base_size}{base font size}

\item{grid}{show grid lines}

\item{serif}{use serif font}
}
\value{
theme object
}
\description{
Similar to the default theme, but with Tufte-esque defaults.
}
