% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_space.R
\name{theme_space}
\alias{theme_space}
\title{Default space theme (light)}
\usage{
theme_space(base_size = 10, grid = TRUE, mono = FALSE)
}
\arguments{
\item{base_size}{base font size}

\item{grid}{show grid lines}

\item{mono}{use monospace font}
}
\value{
theme object
}
\description{
Add to a ggplot2 plot object to apply.
}
