% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wc.R
\name{theme_wc}
\alias{theme_wc}
\title{Classic theme}
\usage{
theme_wc(base_size = 10, grid = FALSE, serif = TRUE)
}
\arguments{
\item{base_size}{base font size}

\item{grid}{show grid lines}

\item{serif}{use serif font}
}
\value{
\code{\link{theme}} object
}
\description{
Similar to the default theme, but with larger serif font and no gridlines.
}
