% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wl.R
\name{theme_wl}
\alias{theme_wl}
\title{Light theme}
\usage{
theme_wl(base_size = 9, grid = TRUE, serif = FALSE)
}
\arguments{
\item{base_size}{base font size}

\item{grid}{show grid lines}

\item{serif}{use serif font}
}
\value{
\code{\link[ggplot2:theme]{ggplot2::theme}} object
}
\description{
Add to a ggplot2 plot object to apply.
}
