% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wt.R
\name{theme_wt}
\alias{theme_wt}
\title{Tufte theme}
\usage{
theme_wt(base_size = 10, grid = FALSE, serif = TRUE)
}
\arguments{
\item{base_size}{base font size}

\item{grid}{show grid lines}

\item{serif}{use serif font}
}
\value{
\code{\link[ggplot2:theme]{ggplot2::theme}} object
}
\description{
Similar to the default theme, but with Tufte-esque defaults.
}
